Clean script checks if files exists and in that case deletes them

+++ MyCode/clean.sh
@@ -1,4 +1,21 @@
-rm nullspacebasis
-rm nullspacebasis.o
-rm gmon.out
-rm .Tpo
+#!/bin/bash
+
+if [ -e nullspacebasis ]
+then
+    rm nullspacebasis
+fi
+
+if [ -e nullspacebasis.o ]
+then
+    rm nullspacebasis.o
+fi
+
+if [ -e gmon.out ]
+then
+    rm gmon.out
+fi
+
+if [ -e .Tpo ]
+then
+    rm .Tpo
+fi
